Refactor utility functions into 'RouterCommon.ts'
